requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreschema==0.0.4
crypto==1.4.1
cryptography==2.7
dj-database-url==0.5.0
Django
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Naked==0.1.31
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
pycrypto==2.6.1
pypi==2.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
shellescape==3.4.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.2
